Chapter 4/restful_python_chapter_04_05/gamesapi/requirements.txt
coverage==4.2
Django==1.10.1
django-crispy-forms==1.6.0
django-filter==0.15.0
django-nose==1.4.4
djangorestframework==3.4.7
nose==1.3.7
